Offline playback with encryption

+++ lrnplayer/src/main/java/io/dotlearn/lrnplayer/listener/OnDownloadProgressListener.kt
@@ -13,6 +13,6 @@ interface OnDownloadProgressListener {
      * @param lrnPlayerView The LRNPlayerView that is downloading the video
      * @param progressPercent The download progress percentage, from 0 to 100.
      */
-    fun onDownloadProgress(lrnPlayerView: LRNPlayerView, progressPercent: Float)
+    fun onDownloadProgress(lrnPlayerView: LRNPlayerView, progressPercent: Int)
 
 }
